# src/main/resources/static/openapi-pokemon.yaml
openapi: 3.1.0
info:
  title: Pokemon API
  version: 1.0.0
  description: API for retrieving Pokemon data
#servers:
#  - url: http://www.pokemondb.com
#    description: Pokemon Database

paths:
  /pokemon/search:
    get:
      summary: Get Pokemon by name
      description: Retrieves a Pokemon's details based on their unique name.
      parameters:
        - name: name
          in: query
          required: true
          description: Name of the Pokemon to retrieve
          schema:
            type: string
            example: Charizard
      responses:
        200:
          description: Successfully retrieved Pokemon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
              examples:
                example-1:
                  $ref: '#/components/schemas/Pokemon'
        404:
          description: Pokemon not found
          content:
            application/json:
              example: null
  /pokemon/add:
    post:
      summary: Adds a new Pokemon
      description: 'Based on the validity of the region(could be name or id)
      it will search the region and add it to the mentioned information for the pokemon'
      requestBody:
        required: true
        description: ''
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PokemonDto"
      responses:
        201:
          description: Pokemon Created Successfully
          content:
            application/json:
              schema:
                type: string
                example: "Charizard was successfully added"

components:
  schemas:
    PokemonType:
      type: string
      description: "Available Pokémon types"
      enum:
        - FIRE
        - WATER
        - GRASS
        - NORMAL
        - ELECTRIC
        - FLYING
        - FIGHTING
        - POISON
        - GROUND
        - PSYCHIC
        - ROCK
        - ICE
        - BUG
        - DRAGON
        - GHOST
        - DARK
        - STEEL
        - FAIRY
        - STELLAR
    PokemonRegion:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: Kanto
        population:
          type: integer
          example: 5000000
    Pokemon:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Charizard
        level:
          type: integer
          example: 36
        ability:
          type: string
          example: Blaze
        types:
          type: array
          items:
            $ref: "#/components/schemas/PokemonType"
          example:
            - FIRE
            - DRAGON
        pokemonRegion:
          $ref: '#/components/schemas/PokemonRegion'
    PokemonDto:
      type: object
      properties:
        name:
          type: string
          example: Charizard
        level:
          type: integer
          example: 10
        ability:
          type: string
          example: Blaze
        pokemonTypes:
          type: array
          items:
            $ref: "#/components/schemas/PokemonType"
          example:
            - FIRE
            - DRAGON
        regionName:
          type: string
          example: Kanto
        regionId:
          type: integer
          example: 1